+++ server/server.js
@@ -42,7 +42,15 @@ app.use(
 	})
 );
 
-app.use(bodyParser.json());
+app.use(bodyParser.json({ limit: '50mb' }));
+app.use(
+	bodyParser.urlencoded({
+		limit: '50mb',
+		extended: true,
+		parameterLimit: 50000,
+	})
+);
+
 app.use('/start', startRouter);
 app.use('/promotions', promotionsRouter);
 app.use('/redemptions', redemptionsRouter);