+++ src/components/Products/index.js
@@ -1,40 +1,17 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import ProductCard from './ProductCard';
 import Spinner from 'react-bootstrap/Spinner';
 import { connect } from 'react-redux';
-import _orderBy from 'lodash.orderby';
-import _isEmpty from 'lodash.isempty';
 import PropTypes from 'prop-types';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
-import ProductsCategoryForm from './ProductsCategoryForm';
 import './style.css';
 import DiscountCarousel from './DiscountCarousel';
 
 const Products = ({ products, fetchingProducts, campaigns }) => {
-	const [filterCategory, setFilterCategory] = useState('');
-
 	useEffect(() => {
 		window.scrollTo(0, 0);
 	}, []);
-
-	let filteredList;
-	if (!fetchingProducts) {
-		switch (filterCategory) {
-			case '':
-			case 'All products':
-				filteredList = products;
-				break;
-			default:
-				filteredList = products.filter(
-					(product) =>
-						!_isEmpty(product.metadata.categories) &&
-						product.metadata.categories.includes(filterCategory)
-				);
-				break;
-		}
-		filteredList = _orderBy(filteredList, ['source_id'], ['asc']);
-	}
 
 	return (
 		<>
@@ -49,27 +26,18 @@ const Products = ({ products, fetchingProducts, campaigns }) => {
 				) : (
 					<>
 						<Row className="products">
-							<Col lg={6}>
-								<h1 className="pageTitle">
-									{filterCategory === '' ? 'Products' : `${filterCategory}`}
-								</h1>
+							<Col lg={12}>
+								<h1 className="pageTitle">Products</h1>
 							</Col>
-							<Col lg={6} md={12} className="productCategoryFormWrapper">
-								<ProductsCategoryForm
-									filterCategory={filterCategory}
-									setFilterCategory={setFilterCategory}
-								/>
-							</Col>
+							{/* <Col lg={6} md={12} className="productCategoryFormWrapper"></Col> */}
 						</Row>
 
 						<Row className="products">
-							{filteredList.map((product) => (
+							{products.map((product) => (
 								<ProductCard
 									key={product.id}
 									product={product}
-									productCampaigns={campaigns.filter(
-										(camp) => camp.metadata.promotion_product === product.name
-									)}
+									campaigns={campaigns}
 								></ProductCard>
 							))}
 						</Row>
@@ -89,7 +57,7 @@ const mapStateToProps = (state) => {
 	};
 };
 
-export default connect(mapStateToProps)(Products);
+export default connect(mapStateToProps)(React.memo(Products));
 
 Products.propTypes = {
 	fetchingProducts: PropTypes.bool.isRequired,